# docs/API_Contract_260714.yaml
openapi: 3.1.0
info:
  title: HUG 안심전세 체인 API Contract
  version: "1.0.0"
  summary: >-
    Frontend / Backend / ML / Blockchain 공통 계약(Single Source of Truth) 문서.
    Backend 구현 문서가 아니며 FastAPI 코드도 아니다.
  description: |
    ## 문서 성격
    본 문서는 HUG × 아이엔 안심전세 체인(HUG 안심전세 체인) 프로젝트의 **Frontend / Backend / ML / Blockchain 공통 계약**이다.
    Backend를 새로 설계하는 문서가 아니라, 이미 확정된 4개 상위 문서(개발설계보고서, 데이터수집 및 API가이드, ML개발가이드,
    Frontend UI/UX 명세서)를 근거로 프론트엔드가 axios 호출을 만들고, 백엔드가 FastAPI를 구현하고, ML Notebook이
    inference endpoint를 구현할 수 있도록 하는 **OpenAPI 3.1 단일 기준선**이다.

    ## 서비스 한 줄 요약
    HUG 안심전세 체인은 위험을 점수화하는 앱이 아니라, 계약 전 위험을 발견하고 위험조건이 실제로 해소되었는지
    증빙과 외부 API로 검증하며, 계약서·검증 이력을 블록체인에 공증하고, 계약 만료 전 D-90 반환계획 확인과
    사고 후 HUG 채권관리 인계까지 하나의 생애주기로 연결하는 주거안전 인프라다.

    ## 선행 문서와 우선순위
    | 영역 | 기준 문서 | 충돌 시 우선순위 |
    |---|---|---|
    | 공통 식별자·상태 | 개발설계보고서_260714_수정보완.docx 10장, 부록 B·C | 최우선 |
    | 외부 API/Mock | 데이터수집_및_API가이드_260714.md | 외부 API 상세 |
    | ML 입출력 계약 | ML개발가이드_260714.md 20장, 부록 C | 모델 Request/Response |
    | 화면·Route·API Mapping | Frontend_UIUX_명세서_260714.md 2·5·8장 | 화면-API 연결 |

    본 문서(API_Contract_260714.yaml)는 위 4개 문서의 API 관련 내용을 통합한 **실행 가능한 계약**이며,
    향후 작성될 Backend_API_명세서_260714.md, Blockchain_설계서_260714.md는 본 계약과 충돌할 수 없다.

    ## 데이터 신뢰성에 대한 명시 (ML개발가이드 2·30장 계승)
    - 정상계약(사고 없이 종료된 계약) 데이터가 없어 "실제 사고확률"을 추정하는 모델은 존재하지 않는다.
    - HUG 사고·경매·배당 데이터 7종은 전량 합성데이터이며, 회수·처리기간 관련 응답은 "합성데이터 기준 패턴"이다.
...

    ## 온체인 금지 원칙 (개발설계보고서 8장)
...

    ## 문서 요약 통계
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /risk/diagnose:
    post:
      operationId: riskDiagnose
      tags: [Property-Risk]
      summary: 통합 위험진단 요청
      description: |
        정규화된 주소와 계약정보를 입력받아 CODEF 등기부(또는 Mock) → 건축물대장 → 실거래가 → 공시가격(Mock)
        → 사업자등록/DART 순으로 외부 API를 오케스트레이션하고(데이터수집_및_API가이드 4장 연동 순서),
        Rule Engine 결과와 ML 위험유사도(RiskPrediction)를 결합해 risk_assessment_id를 생성한다(ML개발가이드 21장 결합 원칙).
        규칙 엔진과 ML 결과가 충돌하면 보수적 결과(고위험)를 우선한다. TEN-02 화면이 진행 상태를 폴링한다.
      x-permission: tenant
      x-mock-supported: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskDiagnoseRequest'
            example:
              property_id: "c1a2b3c4-0000-0000-0000-000000000010"
              deposit: 400000000
              contract_start_date: "2026-08-01"
              contract_end_date: "2028-07-31"
              landlord_type: INDIVIDUAL
              housing_type: MULTI_HOUSEHOLD
      responses:
        "202":
          description: 위험진단 접수(비동기 오케스트레이션 시작). Backend는 이후 GET /risk/{case_id}로 폴링을 유도한다.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          case_id:
                            $ref: '#/components/schemas/UUID'
                          risk_assessment_id:
                            $ref: '#/components/schemas/UUID'
                          api_progress:
                            type: array
                            items:
                              $ref: '#/components/schemas/APIProgressStep'
              example:
                status: Success
                data:
                  case_id: "d2e3f4a5-0000-0000-0000-000000000020"
                  risk_assessment_id: "e3f4a5b6-0000-0000-0000-000000000030"
                  api_progress:
                    - api_name: registry_codef
                      status: Success
                    - api_name: building_registry
                      status: Success
                    - api_name: rtms_price
                      status: Reviewing
                    - api_name: business_status
                      status: Failed
                request_id: "req_20260714_0005"
        "422":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "408":
          $ref: '#/components/responses/Timeout'
  /risk/{case_id}:
    get:
      operationId: getRiskResult
      tags: [Property-Risk]
      summary: 위험진단 결과 조회
      description: |
        case_id 기준 위험진단 결과(위험등급, risk_reasons, 해결가능/해결불가 구분, SHAP 자연어 요약, 데이터 출처,
        조회시각, model_version, ModelResultStatus)를 조회한다. TEN-03 화면의 주 데이터 소스다.
      x-permission: tenant (본인 소유 case_id만)
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/CaseIdPath'
        - $ref: '#/components/parameters/TraceIdHeaderParam'
      responses:
        "200":
          description: 위험진단 결과 조회 성공
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestIdHeader'
            X-Trace-ID:
              $ref: '#/components/headers/TraceIdHeader'
            X-Model-Version:
              $ref: '#/components/headers/ModelVersionHeader'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RiskAssessment'
              example:
                status: Success
                data:
                  risk_assessment_id: "e3f4a5b6-0000-0000-0000-000000000030"
                  case_id: "d2e3f4a5-0000-0000-0000-000000000020"
                  contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                  risk_grade: HIGH
                  risk_reasons: ["근저당설정", "전세가율 높음"]
                  resolvable_risks: ["근저당 말소 예정"]
                  unresolvable_risks: ["건물 구조적 노후"]
                  ml_prediction:
                    risk_grade_suggestion: HIGH
                    similarity_score: 0.82
                    matched_patterns: ["근저당 고비율 사고 패턴"]
                    top_factors:
                      - feature_name: mortgage_ratio
                        contribution: 0.34
                        direction: increase
                    model_version: risk_similarity_v1.0.0
                    model_result_status: Success
                    is_mock: false
                  data_sources: ["codef_registry", "rtms_price", "building_registry"]
                  fetched_at: "2026-07-14T13:30:00+09:00"
                  blockchain_tx_id: null
                  created_at: "2026-07-14T13:31:00+09:00"
                request_id: "req_20260714_0006"
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /evidence-requests:
    get:
      operationId: listEvidenceRequests
      tags: [Evidence-Verification]
      summary: 보완요청 목록 조회
      description: |
        규칙 엔진/아이엔이 생성한 보완요청(EvidenceRequest) 목록을 조회한다. TEN-04 화면(조회 전용)에서 사용하며,
        제출은 임대인 화면(POST /evidence)에서 수행한다.
      x-permission: tenant, landlord
      x-mock-supported: true
      parameters:
        - name: case_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UUID'
        - name: contract_id
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        "200":
          description: 보완요청 목록 조회 성공(0건 포함)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/EvidenceRequest'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
              example:
                status: Success
                data:
                  items:
                    - evidence_request_id: "cc000000-0000-0000-0000-000000000001"
                      contract_id: "f4a5b6c7-0000-0000-0000-000000000040"
                      reason: "근저당 말소 확인"
                      evidence_type: REGISTRY_CANCELLATION_PROOF
                      due_date: "2026-07-20"
                      verification_status: Reviewing
                  pagination: { page: 1, size: 20, total_elements: 1, total_pages: 1 }
                request_id: "req_20260714_0012"
        "401":
          $ref: '#/components/responses/Unauthorized'
  /evidence-requests/{evidence_request_id}:
    get:
      operationId: getEvidenceRequest
      tags: [Evidence-Verification]
      summary: 보완요청 상세 조회
      description: 단건 보완요청 상세(사유, 대상 위험요소, 마감일, 관련 제출 증빙 목록)를 조회한다.
      x-permission: tenant, landlord, advisor
      x-mock-supported: true
      parameters:
        - name: evidence_request_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: 보완요청 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EvidenceRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /evidence:
    post:
      operationId: submitEvidence
      tags: [Evidence-Verification]
      summary: 증빙 제출
      description: |
        임대인이 보완요청에 대한 증빙 파일(등기부 말소 확인서 등)을 업로드한다. 업로드 즉시 document_hash(SHA-256)를
        생성하고 evidence_id를 발급하며 VerificationStatus는 Submitted로 전이한다. LAND-01 화면에서 호출한다.
      x-permission: landlord
      x-mock-supported: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EvidenceSubmitRequest'
      responses:
        "201":
          description: 증빙 제출 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Evidence'
              example:
                status: Success
                data:
                  evidence_id: "dd000000-0000-0000-0000-000000000001"
                  evidence_request_id: "cc000000-0000-0000-0000-000000000001"
                  file_name: "말소확인서.pdf"
                  document_hash: "sha256:9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08"
                  verification_status: Submitted
                  submitted_at: "2026-07-14T14:00:00+09:00"
                request_id: "req_20260714_0013"
        "422":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/Conflict'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /verifications/{verification_id}:
    get:
      operationId: getVerification
      tags: [Evidence-Verification]
      summary: 검증 상태 조회
      description: |
        제출 증빙의 검증 진행 상태(VerificationStatus: Pending~Verified/Rejected/Expired)와 검토 코멘트를 조회한다.
        LAND-02, ADV-02 화면에서 사용한다.
      x-permission: landlord (본인 제출건만), advisor, verifier
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/VerificationIdPath'
      responses:
        "200":
          description: 검증 상태 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Verification'
              example:
                status: Success
                data:
                  verification_id: "ee000000-0000-0000-0000-000000000001"
                  evidence_id: "dd000000-0000-0000-0000-000000000001"
                  verification_status: Reviewing
                  reviewer_comment: null
                  resubmission_required: false
                  blockchain_tx_id: null
                request_id: "req_20260714_0014"
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /verifications/{verification_id}/decision:
    post:
      operationId: decideVerification
      tags: [Evidence-Verification]
      summary: 검증 결정(승인/반려/보류)
      description: |
        아이엔 검증 담당자(advisor, verifier 서브 권한 필요)가 제출 증빙과 외부 API 조회 결과를 대조해
        승인/반려/보류를 결정한다. 승인 시 VerificationStatus=Verified로 전이하고 Blockchain Adapter를 통해
        검증 완료 이벤트를 온체인에 기록한다(개발설계보고서 8장). ADV-02 화면에서 호출한다.
      x-permission: advisor (verifier 서브 권한 필수, 없으면 조회만 허용)
      x-mock-supported: true
      parameters:
        - $ref: '#/components/parameters/VerificationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationDecisionRequest'
            example:
              decision: approve
              reviewer_comment: "등기부 말소 확인서와 CODEF 재조회 결과 일치"
      responses:
        "200":
          description: 검증 결정 처리 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Verification'
              example:
                status: Success
                data:
                  verification_id: "ee000000-0000-0000-0000-000000000001"
                  verification_status: Verified
                  reviewer_comment: "등기부 말소 확인서와 CODEF 재조회 결과 일치"
                  blockchain_tx_id: "ff000000-0000-0000-0000-000000000001"
                request_id: "req_20260714_0015"
        "422":
          $ref: '#/components/responses/ValidationError'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [Admin]
      summary: 사용자·권한 관리 조회
      description: 사용자 역할(tenant/landlord/advisor/hug_admin/verifier) 배정과 계정 상태를 조회한다. ADM-00 화면에서 사용한다.
      x-permission: system_admin
      x-mock-supported: true
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserRole'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        "200":
          description: 사용자 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
              example:
                status: Success
                data:
                  items:
                    - user_id: "8f14e45f-ceea-4b4f-9c1a-000000000001"
                      role: tenant
                      display_name: "홍길동"
                      last_login_at: "2026-07-14T09:12:00+09:00"
                  pagination: { page: 1, size: 20, total_elements: 1, total_pages: 1 }
                request_id: "req_20260714_0025"
        "403":
          $ref: '#/components/responses/Forbidden'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer 토큰. `role` 클레임(tenant/landlord/advisor/hug_admin/system_admin/verifier)을 포함한다.
        `verifier`는 advisor 내 서브 권한으로, `/verifications/{id}/decision` 액션 권한 판정에만 사용한다
        (Frontend_UIUX_명세서 3장, 개발설계보고서 13.5절). 발급/폐기 상세 구현은 Backend_API_명세서_260714.md에서 확정한다.
  parameters:
    CaseIdPath:
      name: case_id
      in: path
      required: true
      description: risk_assessment 조회 단위 케이스 ID
      schema:
        $ref: '#/components/schemas/UUID'
...
    VerificationIdPath:
      name: verification_id
      in: path
      required: true
      description: 검증 결과 식별자
      schema:
        $ref: '#/components/schemas/UUID'
    PageParam:
      name: page
      in: query
      required: false
      description: 페이지 번호(1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      required: false
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TraceIdHeaderParam:
      name: Trace-ID
      in: header
      required: false
      description: 분산 트레이싱용 상관관계 ID(OpenTelemetry trace_id 등).
      schema:
        type: string
...
  headers:
    RequestIdHeader:
      description: 요청-응답을 연결하는 추적 ID. 클라이언트가 보낸 `Request-ID`를 그대로 반환하거나 서버가 생성한다.
      schema:
        type: string
    TraceIdHeader:
      description: 분산 트레이싱 상관관계 ID.
      schema:
        type: string
    ModelVersionHeader:
      description: 응답 생성에 사용된 실제 ML 모델 버전(추론 API에서만 반환).
      schema:
        type: string
...
  responses:
...
    ValidationError:
      description: 요청 값 검증 실패(422). Zod/Pydantic 검증 오류를 필드 단위로 반환한다.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      error_code:
                        type: string
                        example: ERROR-002
                      message:
                        type: string
                        example: "요청 값이 유효하지 않습니다."
                      field_errors:
                        type: array
                        items:
                          $ref: '#/components/schemas/ValidationErrorItem'
          example:
            status: Failed
            error:
              error_code: ERROR-002
              message: "요청 값이 유효하지 않습니다."
              field_errors:
                - field: deposit
                  reason: "0보다 큰 정수여야 합니다."
            request_id: "req_20260714_0100"
    Unauthorized:
      description: 인증 실패(401). 토큰 없음/만료/서명 불일치.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-004, message: "인증 토큰이 유효하지 않습니다." }
            request_id: "req_20260714_0101"
    Forbidden:
      description: 권한 없음(403). 역할(role)이 허용되지 않거나 verifier 서브 권한이 없는 경우.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-005, message: "해당 작업을 수행할 권한이 없습니다." }
            request_id: "req_20260714_0102"
    NotFound:
      description: 리소스 없음(404). 존재하지 않거나 본인 소유가 아닌 리소스에 대한 조회.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-006, message: "요청한 리소스를 찾을 수 없습니다." }
            request_id: "req_20260714_0103"
    Conflict:
      description: 상태 충돌(409). 중복 접수, 상태 전이 규칙 위반 등.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-007, message: "이미 접수된 사건이 존재합니다." }
            request_id: "req_20260714_0104"
    Timeout:
      description: 외부 API/모델 응답 지연(408). 데이터수집_및_API가이드 5장 Fallback Flow 대상.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: Failed
            error: { error_code: ERROR-008, message: "외부 API 응답이 지연되어 재시도가 필요합니다." }
            request_id: "req_20260714_0105"
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: 전역 고유 식별자. 개발설계보고서 10.1절 공통 식별자 16종은 모두 이 포맷을 사용한다.
      example: "8f14e45f-ceea-4b4f-9c1a-000000000001"
...
    Money:
      type: integer
      format: int64
      minimum: 0
      description: KRW 정수 원 단위 금액(개발설계보고서 10.5절). 화면에서는 천 단위 구분을 적용해 표시한다.
      example: 400000000
    Pagination:
      type: object
      description: 목록 API 공통 페이지네이션 응답.
      required: [page, size, total_elements, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        size:
          type: integer
          minimum: 1
          example: 20
        total_elements:
          type: integer
          minimum: 0
          example: 1
        total_pages:
          type: integer
          minimum: 0
          example: 1
...
    SuccessResponse:
      type: object
      description: 모든 성공 응답의 공통 봉투(envelope). data 필드는 각 API별 실제 스키마로 대체된다.
      required: [status, data, request_id]
      properties:
        status:
          type: string
          enum: [Success, Partial]
          example: Success
        data:
          description: 실제 응답 페이로드. 각 API operation에서 구체 스키마로 재정의한다.
          type: object
        request_id:
          type: string
          description: Request-ID 헤더와 동일한 값을 본문에도 포함해 로그 상관관계를 보장한다.
          example: "req_20260714_0001"
...
    RiskGrade:
      type: string
      description: 위험진단 등급(Frontend_UIUX_명세서 5.3절 TEN-03). 규칙 엔진과 ML 위험유사도를 결합한 최종 등급이다.
      enum: [LOW, MEDIUM, HIGH]
      example: HIGH
    UserRole:
      type: string
      description: 사용자 역할(개발설계보고서 13.5절, Frontend_UIUX_명세서 3장). verifier는 advisor 내 서브 권한이다.
      enum: [tenant, landlord, advisor, hug_admin, system_admin, verifier]
      example: tenant
...
    LandlordType:
      type: string
      description: 임대인 유형(개발설계보고서 표 9, ML개발가이드 부록 B landlord_type Feature).
      enum: [INDIVIDUAL, INDIVIDUAL_BUSINESS, CORPORATION]
      example: INDIVIDUAL
    # ======================================================================
    # 3. Entity Schema - 19종 (개발설계보고서 13.3절 핵심 엔터티 + Frontend/HUG 화면에 필요한 보조 엔터티)
    # ======================================================================
    User:
      type: object
      description: 사용자와 역할(개발설계보고서 부록 B user_id).
      required: [user_id, role]
      properties:
        user_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        role:
          $ref: '#/components/schemas/UserRole'
        display_name:
          type: string
          example: "홍길동"
        email:
          type: [string, "null"]
          format: email
        is_active:
          type: boolean
          default: true
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        last_login_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
...
    RiskAssessment:
      type: object
      description: 위험진단 결과 엔터티(개발설계보고서 부록 B risk_assessment_id).
      required: [risk_assessment_id, case_id, risk_grade, risk_reasons]
      properties:
        risk_assessment_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        case_id:
          $ref: '#/components/schemas/UUID'
        contract_id:
          type: [string, "null"]
          format: uuid
        risk_grade:
          $ref: '#/components/schemas/RiskGrade'
        risk_score:
          type: [number, "null"]
          deprecated: true
          description: >-
            초기 버전의 수치형 위험점수(하위호환용). 개발설계보고서 12.4절 "점수보다 근거와 다음 행동 우선" 원칙에 따라
            신규 클라이언트는 risk_grade와 risk_reasons를 사용해야 하며 이 필드는 더 이상 갱신되지 않는다.
        risk_reasons:
          type: array
          items:
            type: string
          example: ["근저당설정", "전세가율 높음"]
        resolvable_risks:
          type: array
          items:
            type: string
          example: ["근저당 말소 예정"]
        unresolvable_risks:
          type: array
          items:
            type: string
          example: ["건물 구조적 노후"]
        ml_prediction:
          $ref: '#/components/schemas/RiskPrediction'
        data_sources:
          type: array
          items:
            type: string
          example: ["codef_registry", "rtms_price"]
        fetched_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
        blockchain_tx_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
    EvidenceRequest:
      type: object
      description: 위험조건 보완요청 엔터티(개발설계보고서 부록 B evidence_request_id).
      required: [evidence_request_id, contract_id, reason, evidence_type]
      properties:
        evidence_request_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        contract_id:
          $ref: '#/components/schemas/UUID'
        risk_assessment_id:
          type: [string, "null"]
          format: uuid
        reason:
          type: string
          example: "근저당 말소 확인"
        evidence_type:
          $ref: '#/components/schemas/EvidenceType'
        due_date:
          type: [string, "null"]
          format: date
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        created_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
    Evidence:
      type: object
      description: 제출 증빙 엔터티(개발설계보고서 부록 B evidence_id).
      required: [evidence_id, evidence_request_id, file_name, document_hash]
      properties:
        evidence_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        evidence_request_id:
          $ref: '#/components/schemas/UUID'
        file_name:
          type: string
          example: "말소확인서.pdf"
        document_hash:
          type: string
          readOnly: true
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        submitted_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
    Verification:
      type: object
      description: 검증 결과 엔터티(개발설계보고서 부록 B verification_id).
      required: [verification_id, evidence_id, verification_status]
      properties:
        verification_id:
          allOf: [{ $ref: '#/components/schemas/UUID' }]
          readOnly: true
        evidence_id:
          $ref: '#/components/schemas/UUID'
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        reviewer_comment:
          type: [string, "null"]
        resubmission_required:
          type: boolean
          default: false
        reviewer_user_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        blockchain_tx_id:
          type: [string, "null"]
          format: uuid
          readOnly: true
        decided_at:
          allOf: [{ $ref: '#/components/schemas/Timestamp' }]
          readOnly: true
...
    APIProgressStep:
      type: object
      description: TEN-02 외부 API 조회 진행 단계 1건.
      required: [api_name, status]
      properties:
        api_name:
          type: string
          enum: [address_normalize, registry_codef, building_registry, rtms_price, official_price, business_status, dart]
        status:
          $ref: '#/components/schemas/APIResultStatus'
    RiskDiagnoseRequest:
      type: object
      required: [property_id, deposit, contract_start_date, contract_end_date, landlord_type, housing_type]
      properties:
        property_id:
          $ref: '#/components/schemas/UUID'
        deposit:
          $ref: '#/components/schemas/Money'
        contract_start_date:
          type: string
          format: date
        contract_end_date:
          type: string
          format: date
        landlord_type:
          $ref: '#/components/schemas/LandlordType'
        housing_type:
          $ref: '#/components/schemas/HousingType'
...
    EvidenceSubmitRequest:
      type: object
      required: [evidence_request_id, file]
      properties:
        evidence_request_id:
          $ref: '#/components/schemas/UUID'
        file:
          type: string
          format: binary
          description: 업로드 파일(등기부 말소 확인서 등). 실제 저장은 Object Storage, 응답에는 document_hash만 포함된다.
    VerificationDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject, hold]
        reviewer_comment:
          type: [string, "null"]
...